+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Change Log
+
+## 1.25.0
+### Breaking
+* Empty action will no longer result in a error
+
+### Changes
+* Added support for using wildcard characters when referring to script
+* Updated circular script detection
+* Updated def dependencies
+
 ## 1.24.1
 * Bug fix for command split with quotes and spaced